refactor(report): extract report assets to files and inline them
