+++ src/cybergraph/assets/report/report.css
@@ -1,0 +1,65 @@
+    :root {
+      color-scheme: light; font-family: Inter, Segoe UI, Arial, sans-serif;
+      --bg: #f5f7fb; --fg: #111827; --panel: #ffffff; --border: #d8e0ea;
+      --muted: #667085; --th: #f0f3f6; --cy-bg: #f8fafc;
+    }
+    @media (prefers-color-scheme: dark) {
+      :root { color-scheme: dark; --bg: #0b1220; --fg: #e5e7eb; --panel: #111827;
+              --border: #243044; --muted: #94a3b8; --th: #1a2334; --cy-bg: #0b1220; }
+    }
+    :root[data-theme="dark"] { color-scheme: dark; --bg: #0b1220; --fg: #e5e7eb; --panel: #111827;
+              --border: #243044; --muted: #94a3b8; --th: #1a2334; --cy-bg: #0b1220; }
+    :root[data-theme="light"] { color-scheme: light; --bg: #f5f7fb; --fg: #111827; --panel: #ffffff;
+              --border: #d8e0ea; --muted: #667085; --th: #f0f3f6; --cy-bg: #f8fafc; }
+    body { margin: 0; background: var(--bg); color: var(--fg); }
+    header { background: radial-gradient(circle at top left, #1d4ed8, #0b1220 42%, #020617); color: white; padding: 32px 36px; position: relative; }
+    main { max-width: 1320px; margin: 0 auto; padding: 28px 24px 48px; }
+    h1 { margin: 0 0 8px; font-size: 30px; }
+    h2 { margin: 28px 0 12px; font-size: 18px; }
+    .muted { color: var(--muted); }
+    .grid { display: grid; grid-template-columns: repeat(auto-fit, minmax(180px, 1fr)); gap: 12px; }
+    .stat, table { background: var(--panel); border: 1px solid var(--border); border-radius: 12px; box-shadow: 0 10px 30px rgba(15, 23, 42, 0.04); }
+    .stat { padding: 16px; }
+    .stat strong { display: block; font-size: 26px; margin-top: 4px; }
+    table { width: 100%; border-collapse: collapse; overflow: hidden; }
+    th, td { padding: 10px 12px; border-bottom: 1px solid var(--border); text-align: left; vertical-align: top; }
+    th { background: var(--th); font-size: 13px; }
+    tr:last-child td { border-bottom: 0; }
+    .pill { display: inline-block; padding: 3px 8px; border-radius: 999px; background: #eef6ff; color: #075985; font-size: 12px; }
+    .path { background: white; border: 1px solid #d0d7de; border-radius: 8px; padding: 12px; margin-bottom: 10px; }
+    .toolbar { display: flex; gap: 10px; flex-wrap: wrap; margin: 0 0 12px; align-items: center; }
+    .toolbar input, .toolbar select, .toolbar button { border: 1px solid #b6c2cf; border-radius: 6px; padding: 9px 10px; background: white; color: #161b22; font-size: 13px; }
+    .toolbar input { min-width: min(360px, 100%); flex: 1; }
+    .toolbar button { cursor: pointer; }
+    .toolbar button:hover { background: #eef2f7; }
+    code { color: #7c2d12; }
+    .explorer { display: grid; grid-template-columns: minmax(0, 1fr) 340px; gap: 14px; }
+    .graph-card, .details {
+      background: var(--panel); border: 1px solid var(--border); border-radius: 16px;
+      box-shadow: 0 18px 45px rgba(15, 23, 42, 0.08);
+    }
+    .graph-card { overflow: hidden; }
+    .graph-head { display: flex; justify-content: space-between; gap: 16px; padding: 14px 16px; border-bottom: 1px solid #e5e7eb; background: linear-gradient(180deg, #ffffff, #f8fafc); }
+    .graph-title { font-weight: 700; }
+    .graph-subtitle { color: #64748b; font-size: 12px; margin-top: 3px; }
+    .graph-badge { align-self: center; padding: 5px 9px; border-radius: 999px; background: #eff6ff; color: #1d4ed8; font-size: 12px; white-space: nowrap; }
+    #cy { height: 640px; background: var(--cy-bg); }
+    .details { padding: 14px; height: 668px; overflow: auto; font-size: 13px; }
+    .details h3 { margin: 0 0 8px; font-size: 15px; }
+    .details .kv { margin: 4px 0; }
+    .details .tag { display: inline-block; padding: 2px 7px; border-radius: 999px; font-size: 11px; color: white; }
+    .legend { display: flex; flex-wrap: wrap; gap: 12px; margin: 8px 0 14px; font-size: 12px; color: #475467; }
+    .legend-item { display: inline-flex; align-items: center; gap: 6px; }
+    .legend-dot { width: 11px; height: 11px; border-radius: 50%; display: inline-block; }
+    .mode-help { margin: -4px 0 12px; color: #64748b; font-size: 13px; }
+    .risk-strip { display: grid; grid-template-columns: repeat(auto-fit, minmax(230px, 1fr)); gap: 10px; margin: 0 0 14px; }
+    .risk-card { border: 1px solid #e2e8f0; background: #fff; border-radius: 12px; padding: 11px; cursor: pointer; }
+    .risk-card:hover { border-color: #93c5fd; box-shadow: 0 8px 24px rgba(37, 99, 235, 0.12); }
+    .risk-card strong { display: block; font-size: 13px; margin-bottom: 5px; }
+    .risk-card span { color: #64748b; font-size: 12px; }
+    .risk-score { float: right; color: #dc2626; font-weight: 700; }
+    .cg-snippet { margin-top: 8px; border: 1px solid var(--border, #d0d7de); border-radius: 8px; overflow: auto; font-family: ui-monospace, SFMono-Regular, Menlo, monospace; font-size: 12px; }
+    .cg-snippet .ln { display: flex; gap: 10px; padding: 1px 8px; white-space: pre; }
+    .cg-snippet .ln .num { color: var(--muted, #94a3b8); user-select: none; min-width: 30px; text-align: right; }
+    .cg-snippet .ln.hl { background: rgba(245, 158, 11, 0.18); }
+    @media (max-width: 820px) { .explorer { grid-template-columns: 1fr; } .details { height: auto; } }
